Added the old DBI migration code from the Hazaar Tool
The Hazaar Tool was going to get this code ported into the new python
version, but seeing as we are now able to execute code in the
application context from the tool, I figured I would add this to the
framework.  This will allow applications to manage their own database
schemas if they choose.

Then, from the CLI tool we can just execute the snapshot() and migrate()
commands via the application runner.