Added method to export models for view
This is the first method where we start blurring the line between models and views.  This will export a model in the HazaarModelView format (label/value) tuple) which can be more easily displayed in a view.

Potentially we will have built-in methods to convert this resulting array into a table or such.