Major DBI Update
This is a massive update to the DBI module.  This update has a lot of clean up and I have removed a lot of redundant unnecessary code.  For exampe I have dumped the schema parameters on most methods.  If you are working with schemas, just specify them in schema.table format.  There was no need to complicate things.

I have also updated the SQLite driver and tested migration on it.  At this point I have successfully migrated the same schema on MySQL and SQLite databases.