Added a whole model label field for HMV rendering
This works with the __toString() method.  When we render a model, if the model has a __toString magic method we call that to render the model instead of rendering it directly.  This will produce the 'value' part.  To get the 'label' part we look for a label() method and get the return value of that.