Minor updates to errorhandlering
Handlering.  haha.  Anyway.

I have moved the application instance set to after config is loaded.  That way if an error occurs while loading the config it doesn't try and use the application in the error output. 

Also tidied up the fatal error page, which is the error shown if no application exists.