Moved script loads to the post() call
This will make pages appear to load faster. It also means that embedding JavaScript into a view will not work if it relies on library (such as jQuery) as it now won't be loaded until after the views are loaded.

This is something I'm going to try and enforce as it really does decrease the page load times and enhances the user experience.  The So we will force all the applications to require scripts and provide mechanisms for them to transfer usable data into the scripts.