Added OAuth2 authorisation bearer support
The request can now be handed an Auth adapter as per Hazaar\Auth\Adater or one of is descendents, such as Hazaar\Auth\Oauth2 or even \Hazaar\Auth\Helper once Oauth2 has been completed.  It will then use the access token in the adapter as the authorisation bearer header for the request.