Added strict model value prepare method
If the value being assigned to a strict model is NOT an array, then we will now call an optional prepare method that allows the data to be prepared into an array.  By default this method does nothing so that the data is then just not used in the model.