Strict model extend tweaks
We can now do a multi-level extend of strict models.  Previously, if the model contained an array it would be merged using array_merge().  Now we actually look to see if there is an "arrayOf" definition and then we loop through the array and if the objects are Strict models themselves, we can call an extend.  Otherwise we array_merge() as before.