HMV: Added support for closures for the edit parameter
Normally edit could be true/false.  It can now also be a closure, meaning that complex criteria can be checked to see if the field should be editable.  A good example of using this is if an 'id' field is null then it can be edited, otherwise it will be displayed, but locked.