Views now handle configs better inc JSON support
This was done because originally I had a restriction in the old INI config files for sending arguments to a view helper when it was being loaded.  So I created a proprietary format that was parsed by the View class if a helper was being loaded.  We still support this, althought it is much improved (now support dot notation) we now also support native object configs such as those available in JSON config files.

fixes #3