HTTP\Request: Now always sending content-length header
Turns out some servers will send back a 411 length required if there is no content-length header and they believe it should exist.  Such as IIS does with a POST even though the POST request has an empty request body.  It is safe to just send the header with a 0 value all the time, I believe.