Removed config processor from the config class
This seemed a bit silly.  A config object should just be a container for configs, nothing more.  The configs are then used elsewhere in the application.  There were only two directive groups that were being processed anyway.  Search paths (which was already being done elsewhere) and PHP settings.  Both of these are now in the application constructor.