Console: Tweaks to loader and PHPINFO section
I realised that the URL was being used to load the module.  This was incorrect and should have been based on the menu items that have been added to groups.  So now the exec method uses the menu array to look up the module to use.

I also added a PHPINFO action with custom styles phpinfo() output.  This is in the 'app' module, but uses the 'sys' url section.