Added OpenSSL file encrypt/decrypt
This is a fully transparent encryption method.  Meaning the file is prefixed with a special code to indicate that it is encrypted and when ANY file is read in the first 3 bytes are checked for the encryption flag.  If the file is encrypted it is automatically decrypted.  When saving the file, encryption is maintained.  Decrypting a file will permenantly turn off its encryption.  This works well with application configs.