Encrypted config files
Added support for encrypted config files.  Encryption cipher is a static public member on the \Hazaar\File class but changing the default is currently not supported.  Configs are loaded BEFORE bootstrap which is the earliest possible place an applcation can change it, and chancing it will cause configs to get all screwed up, so don't do it.  Using the default is the way to go for now anyway.