Updated authorisation method
Previously this would have only worked with OAuth2 adapters.  Now we have getToken() and getTokenType() methods on the base auth adapter so that others can be used as well.  Also means that (and this is what happened here) if I update the storing location of the token, I don't have to update all these sorts of references.