Added authentication check to media controller
It's now possible to enable authentication on the media controller.  If enabled then it will by default lock out all access that is not authenticated.  However if public reads are allowed, then file access is still allowed but access to media commands is not (which in turns protects writes, moves, etc).  This can be enabled using 'cmd' : true in the configs.

Fixes #59