Added better NULL detection/support for non-string fields
What was happening previously was, if an empty string we set on a field and the field was, for example, an integer, then when it was converted to an int using settype(), the integer value would be zero (0).  This isn't what we want.  Now we have an extra check for empty strings and if the target type IS NOT a string, we set it to null.

There is also added support for string nulls.  If the value of the string being set is "null" but the target field IS NOT a string, then it is also converted to a real null.