Fixed HTTP client redirection
We were not parsing the new URI when updating it so it would aways connect to the same server.  Normally this would be OK and is why it's only been noticed now, but if redirecting from http to https the protocol change would be ignored. 

Now we re-parse the URI and all public members are now private to force us to use the right methods.