Enhanced the new root config option
The reason why this was not working before was because the static member we set at instantiation, whereas in a service the root is set later on run.  So we have fixed this by using a filter on the config Map which will update the application root static member whever the config option is set.