Added transparent base64 encoded query strings
When using the hazaar.url() JS method to generate a URL with parameters it is now possible to encode these as a base64 encoded query string automatically.  This is then detected and automatically decoded by the \Hazaar\Applicaiton\Request\HTTP class.  This is useful for obfuscating query strings from nosy users.  It is not secure though and should NOT be considered a type of encryption.  It merely makes it difficult for the average user to read the query string.

Fixes #83