ake() now supports object and properties!!!
Most significant change to ake() in some time.  We now check if the array is actually an object and if so we use property_exists() to see if the key exists as a public property on the object.  The effect is the same rules as with an array but for an object.