Controller\Reponses: Major performance improvements & compression
This is a major update to the program flow for controller responses.  This should increase performance again, especially for cached responses.  As it turns out, for some responses we were still processing the content even though we were returning a 304.  This is now fixed for all response types.

This change also includes the re-integration of the automatic compression mechanism that was ripped out some time ago.  We can now compress/minify/obfuscate JavaScript and CSS on the fly.  This works well with the file modification system so that 304s are returned without compressing the content again.