dataBinder: Added dataBinderValue object
This is a special object that is now used to store all primitive values.  As it is, it will work as previously but it allows the ability to set a label on a value.  This is then used when displaying in HTML (ie: non-inputs).  INPUTs will still use the value.  By default, if no label is set, then the value is displayed so that the previous behavior is consistent.