Model\Strict: Fixed dataBinderValue not converting to array
A stored dataBinderValue is just an array, so it didn't think it needed to convert it.  Now, we look at the array as normal and if the data type is not a "flat" data type we reset the value to NULL.  This will happen when a value data type changes after it has been stored as a dataBinderValue array.