Added some simple security to CDNJS caching
The directory is now created then the link is generated in the initial run method.  Then, when the file is loaded we check if the directory exists and if it does not we do not allow this library to be accessed.  We also check the HTTP_REFERER to make sure it is our application as an added bonus.