Minor changes to add support for PHP built-in web server
My IDE has support for the built-in web server, so I have made a few minor changes here to get that working.  It's a lot slower than developing with IIS Express, but it works.  This also allows developers to use special versions of PHP that are not normally supported by some IDEs.