Major update to View, Layout
I have changed the way the view layouts work in regard to JS requires and CSS links.  This used to be a function of the Hazaar\View class itself.  I have now made this a feature exclusive to the Hazaar\View\Layout class.  Personally, this is the only place I have ever used it and views don't normally load an entire page, they are just part of it.  The Layout class has always been the exception and was intended to be used to load the whole page, including the outer HTML, BODY, etc elements.  So because of this, this is really the only class that needs to support <script> and <link> tags.  On-the-fly scripts can still be included in sub views and they still propogate up to the layout and rendered at the end of the page.