Updated view helper init method
This method now enforces that the view MUST be a Hazaar\View\Layout object.  Init is now only executed when the view is a layout and not when the view is just a normal view.  This prevents helpers from being initialised multiple times as well as trying to set JS requires and CSS links on a view that will never use them.