PHP CLI Startup Detection for rewrite support
The default rewrite setting is now based on the result of php_sapi_name().  If the framework was started from the command line, for example with `composer serve` then the default setting for rewrite is now `false`, meaning that everything should work out of the box regardless of how the app is executed.