Fixed error and debug methods failing when app not initialised
If an error occurrs in the app startup sequence then error capture and output can be a bit messed up because to properly capture error some aspects of the application are checked.  If these elements don't yet exist because the app faied before they were initialised, then more errors are thrown and a possible error loop created.  Error loops are caught, however the error output may not be the actual original cause of the error.  Hopefully this fix will allow errors to be caught and displayed without more being thrown.