Completed initial version of the template engine
This engine uses the SMARTY template format but the code is implemented entirely by me.  Not a single line was "borrowed" from any other engines.  Because of this it is not currently a full implementation so only what I feel I will need initially is working.  This is the usual value replacement feature along with functions such as if, elseif, else, section and sectionelse.

There is also a Hazaar specific built-in function called 'url' which will generate Hazaar application URLs.