Helpers: Updated array_from_dot_notation
Now forcing only a single depth pass over the supplied array as this is how dot notation arrays are supposed to be defined.  If the value is an array then we now leave it along rather than recusing into it.  

NOTE: Merging multi-dimensional arrays and an array in dot notation works perfectly fine now since this update.