jQuery Helper: 'remove' event not optional
This was conflicting with other libraries that already implement it.  So that the event was getting fired for each implementation.  Now we check if it is already implemented and if so do not implement it.  This is a massive hack but it's the only way I can think of doing it.