Action controller return type update
The acton controller will now return a response object of type Json almost always now, unless the response is specifically NULL.  This means that even if you return just "true" or a string, it will be sent as a JSON response.  This is in line with the most common use of the action method return type.