jQuery: Removed the list autoremove function
It was too problematic.  Now, if you want to remove a list item just remove it from the dataBinderArray object and it will update the DOM but removing it from the DOM will no longer update the dataBinderArray object.  This is because when you delete a list that is being displayed the 'remove' events were being fired even though you probably didn't want the item removed and we were just removing the list from display.  In my case this was on a JS page navigation.