HMV: Fixed auto object instantiate messing with some objects
We should ONLY instantiate Strict model objects.  This was previously creating empty Hazaar\Date objects which would render a date of the current timestamp.  Now we only do this if the object is a strict model because that is what we are trying to support by doing this.

Also, date inputs are now outputting the date value in the correct format.