JS: Fixed dataBinder trigger when label has changed
Previously, we only looked at the value and if the value had not changed we skipped the trigger.  However, if only the label had changed on a dataBinderValue then the trigger still would not fire and there may be cases were we want to change the label but not the value.  For example to make use of a null value label being updated.  Now we check both and only if both have not changed do we skip the trigger.