Fixed tag adjacent newlines getting ditched
Turns out this is a thing with the PHP replacement and exec code.  Any newlines adjacent to a PHP end tag "?>" will get ditched.  So for these, we add a space.  This meant doing a bit of an extra pattern match to match the possible \r\n characters after a tag.  If matched, then we add a space then the new line.