HTML Block add/append now flattens
I just tried to use the new children()/find() methods on a block element and was unable to find the element.  The reason was because it was added in an array so the array was stored on the content.  This was fine for output, but searching would not work unless i used find() to recurse through.  Now we flatten arrays on add so that children() will find them.