Re-implemented the GD image resize althorithm
This update will not only handle cropping with alignment (previously just cropped and centered) it should be faster, especially with larger images because we are doing out math on the source image, not the destination image.  Previously we would copy to negative coordinate on the destination image.  Instead now we always write to the same destination coordinates and the source coordinates are altered, meaning we are not wasting processing time on image data that will not end up being visible.