Added response image caching
This is disabled by default because it can potentially take up a lot of space, but it increases performance and reduces server load when requesting resized images quite considerable.  This essentially allows resized images to be shared between sessions/clients.  Meaning an image is resized once, and then subsequent requests are loaded from a cache file.