Cache control tweaks to cached images
Enhanced the method used to cache.  Because we could possibly call multiple functions, we are unable to set the cache key in each function because a resize would not know about a filter call.  Now we just call checkCacheFile() before and give it some parameters to use as a unique ID, namely the request params, this is then used to generate the cache key.