Image cache now using md5 instead of filename
I thought of using an MD5 identifier in the first place but thought it might be too slow.  Turns out that even on a large image (about 1.5MB, which is large for the web) calculating the md5 hash of the file only takes about 5ms.  Which is fair decent.  So now, instead of creating the hash key with the filename we use the md5 of the file, meaning that if the file gets replaced we don't get an old cache file.  It also means that even if the same file is in a multiple locations it will share a cache file.