Updated the Basic and Action controller run methods
Most of the work is now in a shared method called __runAction() on the Basic controller.  This will check the cachedActions and return it if it exists.  This also handles the __default method routing so that when we update that (which is going to be the next commit) we only need to do it in one place.  I also removed and enhanced some stuff which should add a minor performance improvement.