Fixed controller check in Basic controller
In the basic controller we checked if the requested controller is the controller we are in and if not, we make sure we always route to the __default() action.  This stops us from using actions on the default controller directly for controllers that don't exist and are being re-routed.  Previously we relied on the controller name format of  WhateverController.  This caused issues with the Hazaar console because it doesn't conform to that convention.  So I have come up with a better way of handling this.  Now the controller is initialised with the actual name of the controller.  The only thing this affected was the Controller::url() method, which now uses the request object's getControllerName() method to get the current requested controller.