Strict\Models: Updated to support init() calls on all parents
This is a pretty cool update.  So now, if the Strict Model is a extended class, we don't have to re-define all the fields on the parent classes, which is what would have been previously required.  Previously you would only ever extend if you needed the functions on the parent class as extending the fields was pointless as you would have to redefine them all.

Now you can! So if you extend the a strict model class, all the fields defined in the parent classes are maintained and the fields are merged up the chain.  Pretty cool!