Added custom function handlers
It is now possible to register a custom function handler class.  This class will be used for any functions that are not built-in or defined.  So if a function is called the last thing the compiler will do is check if the function handler class has a method of the requested function name and if so, it will set up to use that function to handle the call.

Because of this, we can now pass through the views registered method handler as well so that when we use view templates we get the same custom function handlers as a PHP view which allows us to call code on the controller from the form.