Fixed CDNJS B-Tree database munting issue
Used a MUTEX lock file to allow mutually exclusive access to the CDNJS B-Tree file during a info update process.  This shouldn't impact performance at all as the lock file is only used when an info write is about to take place.  There is also a double check if the file is locked in case it blocked because another process was actually writing the info the current process needed.