Added init responses
If the init method returns a response object, it is sent immediately.  This gives controllers the ability to bail on a request before an action is even executed.  In the past, the init method would throw an exception if it wanted out, but they may return data in an undesirable format.  So now, the init method can return a controller response and send back whatever the heck it wants.