Added global view links/requires caching
Because of this, I have been able to move the CDNJS file caching out of the CDNJS helper and use the new cache mechanism.  For legacy reasons the existing cache_local option in the CDNJS helper options is still supported.  If it is not defined, then the default view cache option is used.

This allows for a new view option of simply "cache" which if TRUE will cache all view references.