Fixed default configs failing to load when no main config file
Configs would FAIL if the main config file did not exist, even if there was adequate defaults supplied.  Now we load configs, if available, and then while loading apply the defaults before checking if there are any options available.  If by then nothing is available, then we consider it a fail.