Fixed console config loading options
Previously, we had to say when to skip using the application config which actually broke a few things because there needs to be an app config for some stuff to work.  So I came up with a better way and that is, once we start loading into the console we reset the app config back to defaults which removes all the loaded helpers and such (which was the problem I was trying to fix) and uses the default config.  This config is what the console is designed to use so everything now works as it should.