Added better support for dynamic child models
If a field is defined as of type 'model' it normally is supposed to have an 'items' element that defines what the model looks like.  Now, we are able to have a field of type 'model' without 'items' and it will create a childModel object with the allow_undefined flag enabled.  This means that more complex data can be stored in the childModel and still benefit from the functionality of strict models.