Added a new field callback called 'prepare'
This callback is always executed, regardless of the exec_filters flag, and is execute BEFORE anything else, giving the model a chance to prepare an individual value before processing it, even on a fresh populate, similar to the prepare() method that is used for preparing the entire data object.

This could be used for field that can accept an update with an integer that is a DB record ID, then the prepare callback uses that ID to lookup a database record and return a row that is then used to populate the field childModel.