Moved get/setResponseType
This was originally on the request, which seemed a bit off, so I have moved this to the Application itself which means that it is more easily accessible.

The main reason is so that we can set the response type more early on in the execution process, in case an error is thrown in something like a controller constructor.