Updated REST controller exception handler
I say updated, but really I removed it.  With the previous updates to the response type mechanism, we can now set the response type a lot earlier on in the execution cycle so there is no need for a custom exception handler as the default Error controller handles JSON responses pretty well.