Updated is_boolean to actually work with booleans!!
Haha.  Ok, so the purpose of is_boolean() was to handle string booleans like yes/not, true/false, on/off, etc.  So the first thing it did was checked if the argument was a string and if not, just return false.  Now we return the result of is_bool() so that is_boolean() can be a more direct replacement of is_bool().