Enhanced shutdown handlers
The shutdown handler will now capture more.  I realised that the best way to check for errors is to first check if headers are sent.  If headers have been sent, then obviously the error is not actually fatal and can be ignored if it exists.  We only want to capture errors that trigger the shutdown abruptly.