Added support for using file_get_contents to download files
I remembered that the reason I used CURL was because it was capable of writing directly to disk, making it possible to download VERY large files.  file_get_contents can not do this, so CURL is the default and FGC is the backup.  It may throw an error if not enough memory is available to download the file.