Added support for boolean defaults to strict models
fixes #165

This allows a default to be set to TRUE for a field that is of a type of class/object.  Setting TRUE will create an empty object.  Alternatively an array may now also be set as the default value.  Previously this would be sent as the first argument of the class constructor call, which worked fine for strict models but not much else.  Now, if the defined class is extends Model\Strict this behavior is maintained, but if not the array is used as an argument list when creating the new object.  This allows other object types to be created as default values in a more efficient manner.