Fixed set default on class as non-array
There is code out there that is setting a default value on an object definition.  For example, the field is of type Hazaar\Date and the default is an integer in epoch.  We now check if the default value is intended for an object but is a non-array and turn it into an array so that it is handled as an arg array.