Helper: array_collate support for key re-indexing
Key re-indexing is where, instead of collating by 2 array elements by key, we use one element as the key, and then the original array as the value.  This allows use to re-index a numeric array and convert it into a associative array.  We may want to do this if we have a numeric array filled with database table rows, for example, and we want to pull out the row ID and use that as the array index.