Smarty: Tweaks and resolved includes
We now include the included files directly.  Previously we would load the files dynamically in case they changed.  That was really only a function of  File\Template\Smarty which cached things, so it has it's own way of dealing with that now that means we can just include the other files directly.