Removed bootstrap simple mode
This was not really needed.  It was only added so that anything that used runStdin didn't try and process routes, so instead of using this I have moved it into run(), which means runStdin() can be executed without it.  The plan is to eventually remove runStdin() as this has been obsoleted by Warlock 3.