Fixed file locks on cache file backend
Since switching to B-Tree we started to have to worry about file locking.  If the cache is opened in two different sections of the application then a lockwait can occur.  Now, we save all the open files, indexed by filename, in a static variable on the file backend class.  If the file is already there, it is open and we just use that.  Otherwise we create a new BTree class.