Updates to controllers and requests
We were using setRequest() and setApplication() methods.  These have been removed and these are set now in the constructor (application) and the initialiser (request).  All controllers have been updated.  This is a bit more work/code but more efficient as we can now ignore the request if the controller does not need it.