Major updates to how URLs are generated
This should be backwards compatible, but it makes more sense now and might be a tiny bit faster.  Basically I have done away with the 'controller' and 'action' stuff and when you generate a URL, you can either supply a path with an array of params, path WITH ? params, or each part of the path as an argument and it will be collapsed with '/' to build the path.  If an argument is an array or stdClass then it is considered a params array AND the last argument.