Updated Request\Http to use `getallheaders()`
This is a weird one.  Basically, `getallheaders()` in PHP is a pain in the ass.  It doesn't normally exist except when using as an apache module.  Then they updated it to exist on CLI.  Now in PHP7.3 they have updated it to exist in FPM.  Thing is, this is where we get all our headers from the web server.  Normally this isn't a problem using `hazaar_request_headers()`, but turns out this function doesn't get the `Authorization` header.  I have switch this over so that if the function exists, it uses it, otherwise if it doesn't a fake one is created that is an alias for `hazaar_request_headers()`