Updated some default hash algos
Did some testing, and for a reasonable length hash (40 characters) SHA1 has become the clear winner in terms of speed, availability and compatibility.  Yes, there are shorter hashes that are faster (although not much) such as CRC32 and ADLER64, but SHA1 has good hash length and will pretty much always be available.  Of course this is all configurable for the pedantic.