Completely changed the way app redirects work
Instead of redirecting with a header() call followed by an exit;, we now simply return a redirect controller response.  It is then up to the calling method to return the response for processing.  This is how it should have been done in the first place as it allows all the shutdown methods to execute correctly.