Model: Change show_hidden to allow def filters
This is 100% backwards compatible.  If the parameter is a boolean (ie: the old method) then it is converted into the new filter format.

The new format allows values to be filtered out by ANY field definition value.  This is handy for allowing multiple outputs on the same data by simply changing the filter.  An example would be adding 'insert=false' which the DBI driver will filter out when it does an INSERT.