Added function to define a new field at runtime
This is actually a bit of a major re-shuffle of code to the definition loader.  Because we do some stuff based on the field type, like model, array, etc.  We need to do that when adding these new fields.  So I have moved that code into the new `add` function and then the load_definition function just calls add for each field.