Object convert to string before HMV output
This will affect everything, but I've done this at the point that objects should be converted to a string ready to be put into a table.  We are outputting HTML here so everything needs to end up as a string.

DBVs were pretty much already working, I think I made changes to the DBV class that fixed this, but this is now definitely making things more consistent.