Added root element support in REST controller
It's now possible to add a root element route on a REST controller.  Previously this would output the describe API output.  It will still do this, but only if no other matching route is found.  This means that if you want to have the describe API output on a GET then have POST actually do something, you can now do that.