Updated File::set_decoded_content()
This method now better supports the features of data URIs.  Namely in that the first section (the header) has data: followed by the mime type, the last section is the encoded content, and everything in between is considered an attribute.  We now also support the 'name' attribute which will set the name of the file automatically.