Added disableCredentialHashing()
This will turn off the multi-algorithm credential hashing algorithm that creates hashed credentials.  This is useful for when the application has performed authentication in some other way and now has the hashed credential, but would like to update the auth system to take over handling authentication.  Execute this method and then use authenticate() by passing the hashed credential to perform authentication.