jQuery dataBinder update now always exec
We used on only trigger the update event on input items.  Now we do it for all elements.  This should be fine as most of the time these elements won't have an update event defined, but now it's possible to put an update event on a non-form input, such as a DIV, and run code when the value is updated.