Updated application link on controllers
Normally the application was set when the controller was created.  This created the problem that creating a controller was not as simple as it needed to be.  I have a situation where I want to create a new controller, then pass it to the application to run.  Once the app sets up things it sets itself as the application on the controller and runs it.  I have updated this to be the standard process for all controllers and the controllers no longer need to be instantiated with an application argument.  This also means the Hazaar\Loader object doesn't need to keep a reference to the application.