Added fix for returning only public class properties
This was a weird one.  If the property is private it should only be accessible internally.  If it is being accessed internally then the property is used directly.  If it is being accessed externally, the __get() magic function is used, which then in turn can access private properties which it happily returned.  This change now checks if the class defined property is public and only returns it if it is.