Cache: Fixed APC backend update after read with keepalive
When we do a read when keepalives are active (the default) we store the value to trigger an 'touch' on close.  If we do a set() after this, the update using the initial value.  So now we just remove any refresh updates when we do a set as we have just reset the TTL anyway.