Email: Improved attachments
Attachments are now stored separately internally, and appended to the content when rendering the email body.  This means that when we call `clear()` we can choose to clear out the attachments as well.  Default behavior is as before where attachments are kept when clear is called, which is called when `send()` is successful.  This means that the default usage would be to send the same file to multiple recipients (using multiple addTo()/send() calls).  However if multiple recipients should be getting different files, then `clear(true)` can be called to remove the previous attachments.