File\Response: Added support for prepending a BOM
When a mime type is used with "charset=" then we look at the charset used and prepend a BOM to the file if a known one does not already exist.  This is useful because BOM would not be added when generating file responses in-memory, and we can automatically prepend the BOM by adding `charset=utf-??` to the setContentType() call.