feat(auth): added getToken()
This function will return the token used to identity a session store.  For JWT, this is the JWT itself.  For serber-side stores it is the session ID.  The format is always an array with a 'token' key.  JWT will also return a 'refresh' key.

This is mainly so that JWT based apps can get the token and use it in the Authorization header instead of as a cookie.  Calling getToken will prevent the JWT cookie from being set automatically.