feat(dbi2): initial commit of DBI2
DBI2 is going to change how drivers are implemented.  The API should remain the same and be 100% compatible with the existing DBI.  In fact, DBI2 will be dropped in as DBI once it is ready and tested as compatible.  I am developing it along side DBI so that I can pilfer code when I need to while still starting with a fresh base structure.

DBI2 will:
* Use traits to define what a driver is capable of.  Already I have PDO and SQL traits.  Others might be JSON/Document, HTTP, etc.  Whatever is required but could result in shared code.
* Allowing implemting non-PDO drivers.
* The first non-pDO driver will be for SurrealDB which is accessed via HTTP/REST.