fix(auth): JWT session storage fixes to data storage
The way I was storing the identity and data was a bit backwards, so this is now changed around so that the JWT token body is stored in the 'data' key and is available as extra auth data.  This means that identity is now stored just as in the other adapters and extracted from 'sub' when the token is decoded.