feat(auth): htpasswd create password file
Rather than throwing an exception on startup if the password file is missing HTPasswd will now carry on without complaint.  Only when the file is being access via a queryAuth/update/delete will it check the file exists and throw an exception.  This means that when calling create we can now create a new passwd file if one does not exist.