fix(btree): added semlocks to BTree file compact
There was an instance where a BTree file could attempt to compact multiple times concurrently.  This change creates a semlock file that causes one process to hold off if it happens to execute a compact at the exact same time.  This file is also checked when the BTree object is created to prevent use when another process is in the middle of compacting the file.