fix(btree): more file lock enhancements
I found that when calling 'set', it was possible to obtain a LOCK_EX but then during the update another LOCK_SH was obtained during a lookup.  At the end of the lookup the lock was released, however the 'set' had not completed.  This is what I think was causing the file corruption.

To fix this, all locks are requested through a BTree file locked and there are now 3 states of the file lock.
* LOCK_OK means the file was locked successfully.  The calling function then knows it needs to unlock the file.
* LOCK_NO means the lock failed.  This is when calling File::lock() returns false.
* LOCK_NA means the lock was uneccessary and the calling function should not unlock the file.