fix(ratelimiter): file backend garbage collection issue
The garbage collector would remove entries at cleanup time that should not be removed.  The reason was that it assumed the windowLength was the same for all log entries, but some rules can be used with different windowLength values.

To fix this, we now store the windowLength in the log container and use that to calculate the expiration time.  Only if this window value does not exist do we use the internal windowLength value.