fix(route): Action args defined in route callable
It is now possible to define default action args in the route callable.  This means you can declare a route without any arguments that executes an action with require parameters by defining the argument values in the route file.

Such as:

```php
Router::get('/test/{string:name}', [Index::class, 'test']);
Router::get('/test', [Index::class, 'test']);
```

In this example, the first route takes the argument from the route URL.  The second one uses the same action but defines the argument in the callable.